+++ src/http_ui.h
@@ -1,7 +1,0 @@
-/* SPDX-License-Identifier: GPL-2.0-only OR GPL-3.0-only */
-/* Copyright (c) 2021 - 2022 Gavin Henry <ghenry@sentrypeer.org> */
-
-#ifndef SENTRYPEER_HTTP_UI_H
-#define SENTRYPEER_HTTP_UI_H 1
-
-#endif //SENTRYPEER_HTTP_UI_H